Customisable welcome card
Allow the user to create their own welcome card (the first card that is
posted to the user's timeline when they log into the glassware). To create
a custom card, the user should create a service called WelcomeService,
with a method

TimelineItem makeWelcomeCard()

This method should create and return a timeline item that will be posted
to the timeline by the authorisationService. If no such service or method
exists, the default "Welcome to APPNAME" card will be used. If a method
exists but does not return a TimelineItem, a runtime error will be
thrown