Support for image attachments
Added feature for rudimentary support of image attachments. If the user
performs a 'share' action on a card, and the card contains image
attachments, they are retrieved, and stored as a BufferedImage on the
params.attachments list passed to the messageHandlerService. Attachments
for all other actions are not yet supported, and behaviour when the
attachment is not an image (readable by javax.imageio.ImageIO) is not yet
specified.