Application image now ships with Glass Apps
Previously, you had to specify a URL for your application image in the
config. This is the image used for the application in the contact list in
Glass. Now you simply include an image file called contact.png in your
"web-app/images" directory. We might want to change the name & file type
later but for now this is a quicker way to get things up and running.