Fix to allow connectivity with real Google Mirror service.
TEST_ROOT_URL and TEST_SERVICE_PATH were statically typed as Strings. This
leads to odd behaviour in ConfigObject where they are both non-null (as
ConfigObject creates them automatically when they are referenced) and
non-false according to Groovy truth. Basically the ConfigObject toString
is passed into the string variable. Removing the static typing is enough
to make it behave as expected.